renamed viz_libs to viz; move viz functions into util
